Refactor by using extract method,move method,pulldown  method,extract class
